fix: 게시글 조회 쿼리 N+1 문제 수정 및 최적화

+++ server/src/main/java/teamdropin/server/domain/post/dto/GetAllPostResponseDto.java
@@ -13,11 +13,11 @@ public class GetAllPostResponseDto {
     private Long id;
     private String title;
     private String body;
-    private int viewCount;
+    private long viewCount;
     private Category category;
-    private int likeCount;
+    private long likeCount;
     private String nickname;
-    private int commentCount;
+    private long commentCount;
     private LocalDateTime createdDate;
     private String profileImageUrl;
